+++ libs/ui-toolkit/src/components/trading-multi-select/index.ts
@@ -1,0 +1,1 @@
+export * from './multi-select';
